FTPT for tables working

+++ api/src/main/java/ar/edu/itba/pod/QueryService.java
@@ -7,9 +7,10 @@ import java.rmi.Remote;
 import java.rmi.RemoteException;
 import java.util.List;
 import java.util.Map;
+import java.util.concurrent.atomic.AtomicLong;
 
 public interface QueryService extends Remote {
     List<Map.Entry<Party,Long>> getNationalResults() throws RemoteException, ElectionNotStartedException;
     List<Map.Entry<Party,Long>> getProvinceResults(Province province) throws RemoteException, ElectionNotStartedException;
-    List<Map.Entry<Party,Long>> getTableResults(Integer tableID) throws RemoteException, ElectionNotStartedException;
+    List<Map.Entry<Party, Double>> getTableResults(Integer tableID) throws RemoteException, ElectionNotStartedException;
 }
